finish refactoring the api
